refactor lstm code for quantization
